monitoring/backup-monitor/requirements.txt
prometheus-client==0.19.0
psycopg2-binary==2.9.9
requests==2.31.0
